+++ NEWS.md
@@ -6,6 +6,12 @@
    
 ## Minor changes
 
+- `ggboxplot()`, `ggviolin()`, `ggdotplot()` and `ggstripchart()` can now handle one single numeric vector.
+
+```
+# Example
+ggboxplot(iris$Sepal.Length)
+```
    
 ## Bug fixes